refactor: Step 1 - extract danmaku skip to src/content/danmakuSkip.ts
Move checkDanmaku, danmakuForSkip, and related state to a dedicated
module. Use dependency injection for cross-module function calls.
